Remove flag icons and use inline icons + dom based creds
flag-icon-css was too much hassle, so we're hosting all the flag svgs
up on media.g.com and referencing them via image. It's cleaner for
everybody.

the caution icons were never going to work, so turn them into react
components and embed directly. they're happy now

It looks like we want to pass in gapi creds through the dom, so
get some globals and pass those into the api when making the requests