Rename nationalities to countries and move to pkg
Instead of making 3 large requests to the api every time a
trip page is hit, let's use a static dump that's easy to
update. G-Countries was borne of this idea. If the api data
ever updates, we just publish a new version, it gets scraped
and we reference this new version.

Also, with this, the nationalities was renamed to
countries, as a more re-usable term.